refactor: hoc 을 hook으로 전환
